Quiz, CourseUsers functionality fixed, UI improvement

+++ Models/UserPoints.cs
@@ -5,6 +5,6 @@ namespace LMS.Models
     public class UserPoints
     {
         public User UserEntity { get; set; }
-        public int? AchievedPoints { get; set; }
+        public decimal? AchievedPoints { get; set; }
     }
 }
